add break to grid search

+++ model_selection.py
@@ -36,9 +36,9 @@ parser.add_argument('-gpu', type=int, default=0,
 parser.add_argument('-clean',action='store_true', default=False, help='set if to clean the files[default:False]')
 parser.add_argument('-train-file', type=str, default='all.tsv', help='train data tsv name [default:all.tsv]')
 parser.add_argument('-label-file', type=str, default='labels.txt', help='example for each classes [default:labels.txt]')
-parser.add_argument('-predict-result', action='store_true', help='if to predict result on vp task')
+parser.add_argument('-predict-result', action='store_true', default=False, help='if to predict result on vp task')
 parser.add_argument('-result-file', type=str, default='./tmp/result.csv', help='path to store evaluation result [default:./tmp/result.csv]')
-
+parser.add_argument('-oneset', action='store_true', default=False, help='set true to test just one hyperparameter set, instead of grid search')
 
 args = parser.parse_args()
 DATA_BASE_PATH = './data'
@@ -98,7 +98,7 @@ def generate_command(max_seq_length, lr, batch, epoch, datapath, trial_identifie
     command += ' --max_seq_length=' + str(max_seq_length) + ' --train_batch_size='+str(batch)+' --learning_rate='+str(lr)
     command += ' --num_train_epochs='+str(epoch) + ' --output_dir='+output_dir
     if predict:
-        command += '--do_predict=true'
+        command += ' --do_predict=true'
     return command
 
 def get_acc(output_dir):
@@ -127,12 +127,13 @@ for max_seq_length in range(args.seqlen_low, args.seqlen_high+1, args.seqlen_ste
                     command = generate_command(max_seq_length,lr, batch_size, epoch,data_path,trial_id, current_trial_dir, args.predict_result)
                     try:
                         os.system(command)
-                        acc_sum += get_acc(current_trial_dir)
+                        #acc_sum += get_acc(current_trial_dir)
                     except tf.errors.ResourceExhaustedError:
                         acc_sum += -99999
                     trial_id += 1
                     # remove bert checkpoint etc.
-                    shutil.rmtree(current_trial_dir)
+                    if args.clean:
+                        shutil.rmtree(current_trial_dir)
 
                 acc = acc_sum/float(len(validation_dirs))
                 result.append([max_seq_length, lr, batch_size, epoch,acc])
@@ -140,6 +141,8 @@ for max_seq_length in range(args.seqlen_low, args.seqlen_high+1, args.seqlen_ste
                 df = pd.DataFrame(data=result_, columns=['max_seq_len', 'lr', 'batch_size', 'epoch', 'acc'])
                 df.to_csv(args.result_file)
                 trial_id+=1
+                if args.oneset:
+                    quit()
             trial_id +=1
         trial_id +=1
     trial_id +=1
